changed the arraylist datatype from integer to double

+++ Calculator Assignment/src/com/prajwal/Multiplication.java
@@ -3,12 +3,12 @@ package com.prajwal;
 import java.util.ArrayList;
 
 public class Multiplication {
-    ArrayList<Integer>numbers=new ArrayList<Integer>();
-    public Multiplication(ArrayList<Integer>values){
+    ArrayList<Double>numbers=new ArrayList<Double>();
+    public Multiplication(ArrayList<Double>values){
         numbers=values;
     }
-    public int multiply(){
-        int temp=1;
+    public double multiply(){
+        double temp=1;
         for(int i=0;i<numbers.size();i++){
             temp*= numbers.get(i);
         }
